LayoutTests/gamepad/gamepad-event-expected.txt

Tests the GamepadEvent constructor

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS new GamepadEvent() threw exception TypeError: Not enough arguments.
PASS event = new GamepadEvent('test') did not throw exception.
PASS event.type is "test"
PASS event.gamepad is null
PASS event.bubbles is false
PASS event.cancelable is false
PASS event.composed is false
PASS event = new GamepadEvent('test', {'bubbles': true, 'cancelable': true, 'composed': true }) did not throw exception.
PASS event.gamepad is null
PASS event.bubbles is true
PASS event.cancelable is true
PASS event.composed is true
PASS navigator.getGamepads is defined.
PASS successfullyParsed is true

TEST COMPLETE
